Changes to partitioning.
Previously, hash partitioner using default Object.hashCode was used.  That
hashCode is reference-specific, so it wasn't even putting equal keys
on the same reduce parition.  Now, the Edge object has a hashCode method
that applies an array hash to dst.  This makes it highly likely that
map outputs with the same key come from the same map task, shifting almost
all the reduce work to the combiner.

I wrote a custom partitioner, but it is not clear that it is working (or
even being called).